Added: ImageSheet class. Updated: class hierarchy changed dramatically. Image::Sprite and ImageSheet::Animation::AnimatedSprite. Updated: ConfigFile class now does additional checks when using setValue so that you can check if the file contents have changed, preventing redundent writes.

+++ branches/PenjinNew/Panel.cpp
@@ -115,7 +115,7 @@ void Panel::update()
         widgets.at(i)->setActive(false);
     }
 
-    if(selection > -1 && selection < widgets.size())
+    if(selection > -1 && selection < (int)widgets.size())
         widgets.at(selection)->setSelected(true);
     else
         selection = -1;
@@ -123,7 +123,7 @@ void Panel::update()
 
 void Panel::selectionConfirm()
 {
-    if(selection > -1 && selection < widgets.size())
+    if(selection > -1 && selection < (int)widgets.size())
         widgets.at(selection)->setActive(true);
     else
         selection = -1;
